% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellType.R
\docType{class}
\name{CellType}
\alias{CellType}
\alias{CellType-class}
\title{The CellType class}
\description{
Classifies cells using a criterion function. Classifies cells via a user-defined gating function.
The gating function accepts as input the entire matrix of expression data from a \code{CellDataSet},
and return TRUE or FALSE for each cell in it, depending on whether each meets the criteria in the gating function.
}
\section{Slots}{

\describe{
  \item{\code{classify_func}:}{A function that accepts a matrix of expression values as input, and returns a logical vector (of length equal to the number of columns in the matrix) as output}
}
}

