% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_conversion.R
\name{exportCDS}
\alias{exportCDS}
\title{Export a monocle CellDataSet object to a Seurat object}
\usage{
exportCDS(monocle_cds, export_all = FALSE)
}
\arguments{
\item{monocle_cds}{the Monocle CellDataSet you would like to export into a Seurat object}

\item{export_all}{Whether or not to export all the slots in Monocle and keep in the Seurat object. Default is FALSE (or only keep
minimal dataset). If export_all is setted to be true, the original monocle CellDataSet object will be keeped in the Seurat object too.}
}
\value{
a new Seurat object.
}
\description{
Export a monocle CellDataSet object to a Seurat object
}
\examples{
\dontrun{
lung <- load_lung()
seurat_lung <- exportCDS(lung)
seurat_lung_all <- exportCDS(lung, export_all = T)
}
}
