\docType{HSMM}
\name{HSMM}
\alias{HSMM}
\title{The HSMM single-cell RNA-Seq timeseries from Trapnell, Cacchiarelli et al.}
\description{
The HSMM single-cell RNA-Seq timeseries from Trapnell,
Cacchiarelli et al.
}
\author{
Cole Trapnell \email{cole@broadinstitute.org}
}
\references{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52529}
}
\keyword{data}

