% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellDataSet.R
\name{reducedDimW}
\alias{reducedDimW}
\title{Get the whitened expression values for a CellDataSet.}
\usage{
reducedDimW(cds)
}
\arguments{
\item{cds}{A CellDataSet object.}
}
\value{
A matrix, where each row is a set of whitened expression values for a feature and columns are cells.
}
\description{
Retrieves the expression values for each cell (as a matrix) after whitening
during dimensionality reduction.
}
\examples{
\dontrun{
W <- reducedDimW(HSMM)
}
}
