% Generated by roxygen2 (4.0.1): do not edit by hand
\name{selectNegentropyGenes}
\alias{selectNegentropyGenes}
\title{Filter genes with extremely high or low negentropy}
\usage{
selectNegentropyGenes(cds, lower_negentropy_bound = "0\%",
  upper_negentropy_bound = "99\%", expression_lower_thresh = 0.1,
  expression_upper_thresh = Inf)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{lower_negentropy_bound}{the centile below which to exclude to genes}

\item{upper_negentropy_bound}{the centile above which to exclude to genes}

\item{expression_lower_thresh}{the expression level below which to exclude genes used to determine negentropy}

\item{expression_upper_thresh}{the expression level above which to exclude genes used to determine negentropy}
}
\value{
a vector of gene names
}
\description{
Filter genes with extremely high or low negentropy
}
\examples{
\dontrun{
reasonableNegentropy <- selectNegentropyGenes(HSMM, "07\%", "95\%", 1, 100)
}
}

