% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{selectTopMarkers}
\alias{selectTopMarkers}
\title{Select the most cell type specific markers}
\usage{
selectTopMarkers(marker_specificities, num_markers = 10)
}
\arguments{
\item{marker_specificities}{The dataframe of specificity results produced by \code{\link{calculateMarkerSpecificity}()}}

\item{num_markers}{The number of markers that will be shown for each cell type}
}
\value{
A data frame of specificity results
}
\description{
This is a handy wrapper function around dplyr's top_n function to extract
the most specific genes for each cell type. Convenient, for example, for
selecting a balanced set of genes to be used in semi-supervised clustering
or ordering.
}
