% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimateSizeFactorsForMatrix}
\alias{estimateSizeFactorsForMatrix}
\title{Function to calculate the size factor for the single-cell RNA-seq data}
\usage{
estimateSizeFactorsForMatrix(
  counts,
  locfunc = median,
  round_exprs = TRUE,
  method = "mean-geometric-mean-total"
)
}
\arguments{
\item{counts}{The matrix for the gene expression data, either read counts or FPKM values or transcript counts}

\item{locfunc}{The location function used to find the representive value}

\item{round_exprs}{A logic flag to determine whether or not the expression value should be rounded}

\item{method}{A character to specify the size factor calculation appraoches. It can be either "mean-geometric-mean-total" (default),
"weighted-median", "median-geometric-mean", "median", "mode", "geometric-mean-total".}
}
\description{
@importFrom stats median
}
