% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_conversion.R
\name{importCDS}
\alias{importCDS}
\title{Import a Seurat object and convert it to a monocle CellDataSet object}
\usage{
importCDS(srt, import_all = FALSE)
}
\arguments{
\item{srt}{the Seurat object you would like to convert into a monocle CellDataSet object}

\item{import_all}{Whether or not to import all the slots in Seurat.
Default is FALSE (or only keep minimal dataset).}
}
\value{
a new monocle CellDataSet object converted from a Seurat object.
}
\description{
Import a Seurat object and convert it to a monocle CellDataSet object
}
\examples{
\dontrun{
lung <- load_lung()
seurat_lung <- exportCDS(lung)
seurat_lung_all <- exportCDS(lung, export_all = T)

importCDS(seurat_lung)
importCDS(seurat_lung, import_all = T)
importCDS(seurat_lung_all)
importCDS(seurat_lung_all, import_all = T)
}
}
