% Generated by roxygen2 (4.0.0): do not edit by hand
\name{minSpanningTree}
\alias{minSpanningTree}
\title{Retrieves the minimum spanning tree generated by Monocle during cell ordering.}
\usage{
minSpanningTree(cds)
}
\arguments{
\item{cds}{expression data matrix for an experiment}
}
\value{
An igraph object representing the CellDataSet's minimum spanning tree.
}
\description{
Retrieves the minimum spanning tree generated by Monocle during cell ordering.
}
\examples{
data(HSMM)
T <- minSpanningTree(HSMM)
}

